requirements.txt
# ── LLM Clients ──────────────────────────────────────────────────────────────
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# ── Browser Automation (Gemini Web Skill / Screenshot Skill) ─────────────────
playwright>=1.48.0

# ── Image Processing (crop_resize / text_overlay / watermark Skills) ─────────
Pillow>=11.0.0
# opencv-python-headless>=4.10.0  # 可选，watermark Skill mosaic 模式增强版

# ── Video Processing (video_clip Skill) ──────────────────────────────────────
moviepy>=1.0.3

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
aiohttp>=3.10.0
aiofiles>=24.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
